Initial network config monitoring code.
There are two parts to this: first we query the current addresses, then
we subscribe to changes.

There's a small race here, if the addresses change between us doing the
initial query and the start of the monitoring. It's not clear how this
can be fixed, and it's unlikely to every happen in any case.